Create general layout, introductory mathematics
I've tried to incorporate the various areas shown in the [mathematics roadmap](https://github.com/TalalAlrawajfeh/mathematics-roadmap), as well as everything required by the 2015 CUPM Curriculum. I'm not sure if "Introduction to Mathematical Thinking" would be considered enough to cover the history of mathematics requirements that the CUPM curriculum puts forth. 

These suggestions are just suggestions, and most likely, as we find courses and textbooks to cover them, the true structure of the repository will change.