Update README.MD
The original description for this course may have led to confusion about the nature of Analysis. Analysis is a branch of mathematics that deals with the study of limits, properties of functions, and the infinite. While it does make use of formal proof, it is not to be confused with metamathematics which is the study of mathematical reasoning and foundations.